client/requirements.txt
# Client dependencies
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-socketio>=5.3.0
eventlet>=0.33.3
ollama==0.1.6
flask-limiter>=3.5.0
